+++ crates/core-new/src/lib.rs
@@ -1,6 +1,5 @@
 #![deny(unused_crate_dependencies)]
 
 pub mod cairo_runner;
-pub mod input_generator;
 pub mod pre_processor;
 pub mod processor;
