+++ crates/core-new/src/processor/mod.rs
@@ -1,37 +1,55 @@
 use anyhow::Result;
+use input::ProcessorInput;
 
-use crate::{
-    cairo_runner::run::{RunResult, Runner},
-    pre_processor::PreProcessResult,
-};
+use crate::cairo_runner::run::{RunResult, Runner};
 
+pub mod input;
+
+/*
+    Processor is reponsible for running the module.
+    This run is sound execution of the module.
+    This will be most abstract layer of the processor.
+*/
 pub struct Processor {
-    fetch_points: Vec<String>,
-    module_hash: String,
-    module_bytes: Vec<u8>,
+    runner: Runner,
+}
+
+impl Default for Processor {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 impl Processor {
-    pub fn new(pre_processed: PreProcessResult) -> Self {
-        // get module bytes from module hash
-        // rearrange fetch points considering duplicated points
-        let module_bytes = vec![];
-        Self {
-            fetch_points: pre_processed.fetch_points,
-            module_hash: pre_processed.module_hash,
-            module_bytes,
-        }
+    pub fn new() -> Self {
+        let runner = Runner::new();
+        Self { runner }
     }
 
-    pub fn process(&self) -> Result<RunResult> {
+    pub fn process(
+        &self,
+        module_input: Vec<u8>,
+        module_hash: String,
+        fetch_points: Vec<String>,
+    ) -> Result<RunResult> {
         // generate input file from fetch points
         // 1. fetch proofs from provider by using fetch points
-
+        let proofs = vec![];
         // 2. generate input struct with proofs and module bytes
-
+        let input = self.generate_input(proofs, module_input, module_hash);
         // 3. pass the input file to the runner
-        let runner = Runner::new();
-        let input_bytes = vec![];
-        runner.run(input_bytes)
+        let input_bytes = input.to_bytes();
+        self.runner.run(input_bytes)
+    }
+
+    pub fn generate_input(
+        &self,
+        proofs: Vec<String>,
+        module_input: Vec<u8>,
+        module_hash: String,
+    ) -> ProcessorInput {
+        // TODO: get module bytes from registry by module_hash
+        let module_bytes = vec![];
+        ProcessorInput::new(module_bytes, module_input, proofs)
     }
 }
