+++ crates/provider/src/lib.rs
@@ -1,3 +1,4 @@
 #![deny(unused_crate_dependencies)]
 
 pub mod evm;
+pub mod provider;
